# Conflicts:
#	energymanager/src/main/java/com/group1/energymanager/DTOs/TransactionDTO.java
#	energymanager/src/main/java/com/group1/energymanager/controller/TransactionController.java
#	energymanager/src/main/java/com/group1/energymanager/controller/UserController.java
#	energymanager/src/main/java/com/group1/energymanager/exceptions/PacketNotFoundException.java
#	energymanager/src/main/java/com/group1/energymanager/exceptions/UserNotFoundException.java
#	energymanager/src/main/java/com/group1/energymanager/model/User.java
#	energymanager/src/main/java/com/group1/energymanager/request/TransactionRequest.java
#	energymanager/src/main/java/com/group1/energymanager/request/UserRequest.java
#	energymanager/src/main/java/com/group1/energymanager/response/DeleteUserResponse.java
#	energymanager/src/main/java/com/group1/energymanager/response/DepositOnWalletResponse.java
#	energymanager/src/main/java/com/group1/energymanager/response/ListTransactionResponse.java
#	energymanager/src/main/java/com/group1/energymanager/response/TransactionResponse.java
#	energymanager/src/main/java/com/group1/energymanager/response/UpdateUserResponse.java
#	energymanager/src/main/java/com/group1/energymanager/service/TransactionService.java
#	energymanager/src/main/java/com/group1/energymanager/service/UserService.java
#	energymanager/src/main/resources/application.properties

+++ energymanager/src/main/java/com/group1/energymanager/response/DepositOnWalletResponse.java
@@ -4,7 +4,7 @@ public class DepositOnWalletResponse {
 
     private BaseResponse result;
     private String id;
-    private Double wallet;
+    private float wallet;
 
     public BaseResponse getResult() {
         return result;
@@ -22,11 +22,11 @@ public class DepositOnWalletResponse {
         this.id = id;
     }
 
-    public Double getWallet() {
+    public float getWallet() {
         return wallet;
     }
 
-    public void setWallet(Double wallet) {
+    public void setWallet(float wallet) {
         this.wallet = wallet;
     }
 }
